requirements.txt
aiogram==3.15.0
asyncpg==0.30.0
aiohttp==3.10.11
python-dotenv==1.1.0
